requirements.txt
pandas==1.4.3
Pillow==9.1.1
plotly==5.9.0
streamlit==1.10.0
